set initial state of video player
so it's clear the user has to click play in order for the radio to do something.
previously the video just started playing so users thought there was no sound. even though the sound comes from the music player (not the video). and we can't autoplay onload due to browser restrictions. so, setting the video to off by default just makes it more intuitive